requirements.txt
annotated-types==0.7.0
anyio==4.6.0
asgiref==3.8.1
black==24.8.0
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
distro==1.9.0
Django==5.1.1
django-cors-headers==4.4.0
django-debug-toolbar==4.4.6
django-environ==0.11.2
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
environ==1.0
exceptiongroup==1.2.2
filelock==3.16.1
fsspec==2024.9.0
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
huggingface-hub==0.25.2
idna==3.10
Jinja2==3.1.4
jiter==0.5.0
MarkupSafe==3.0.1
mpmath==1.3.0
mypy-extensions==1.0.0
networkx==3.4.1
numpy==2.1.2
openai==1.51.0
packaging==24.1
pathspec==0.12.1
platformdirs==4.3.6
psycopg2-binary==2.9.9
pydantic==2.9.2
pydantic_core==2.23.4
PyJWT==2.9.0
python-dotenv==1.0.1
PyYAML==6.0.2
regex==2024.9.11
requests==2.32.3
safetensors==0.4.5
schedule==1.2.2
sniffio==1.3.1
sqlparse==0.5.1
sympy==1.13.3
tokenizers==0.20.1
tomli==2.0.1
torch==2.4.1
tqdm==4.66.5
transformers==4.45.2
typing_extensions==4.12.2
urllib3==2.2.3
